+++ mongodb/collection_operator.go
@@ -1,0 +1,70 @@
+package mongodb
+
+import (
+	"context"
+
+	"github.com/Becklyn/clerk"
+	"go.mongodb.org/mongo-driver/bson"
+	"go.mongodb.org/mongo-driver/mongo"
+	"go.mongodb.org/mongo-driver/mongo/options"
+)
+
+type MongodbCollectionOperator struct {
+	connection *MongodbConnection
+	client     *mongo.Client
+}
+
+func NewMongoCollectionOperator(connection *MongodbConnection) *MongodbCollectionOperator {
+	return &MongodbCollectionOperator{
+		connection: connection,
+		client:     connection.client,
+	}
+}
+
+func (o *MongodbCollectionOperator) Rename(
+	ctx context.Context,
+	collection *clerk.Collection,
+	renameTo string,
+	drop bool,
+) (*clerk.Collection, error) {
+	handler := func(ctx context.Context) error {
+		cursor, err := o.client.
+			Database(collection.Database).
+			Collection(collection.Name).
+			Find(ctx, bson.D{}, options.Find())
+		if err != nil {
+			return err
+		}
+		var results []interface{}
+		if err := cursor.All(ctx, &results); err != nil {
+			return err
+		}
+
+		_, err = o.client.Database(collection.Database).
+			Collection(renameTo).
+			InsertMany(ctx, results)
+		if err != nil {
+			return err
+		}
+
+		if drop {
+			return o.Drop(ctx, collection)
+		}
+		return nil
+	}
+
+	if err := o.connection.WithTransaction(ctx, handler); err != nil {
+		return nil, err
+	}
+	return clerk.NewCollectionWithDatabase(collection.Database, renameTo), nil
+}
+
+func (o *MongodbCollectionOperator) Drop(
+	ctx context.Context,
+	collection *clerk.Collection,
+) error {
+	return o.client.
+		Database(collection.Database).
+		Collection(collection.Name).
+		Drop(ctx)
+}